feat: add typescript #84
