+++ web/index.html
@@ -23,7 +23,7 @@
     </head>
     <body>
         <div id="app"></div>
-        <!-- <script src="./wasm_exec.js"></script> -->
-        <script type="module" src="/src/main.js"></script>
+        <script src="/wasm_exec.js"></script>
+        <script type="module" src="/src/main.ts"></script>
     </body>
 </html>
