+++ web/package.json
@@ -5,7 +5,9 @@
     "scripts": {
         "dev": "vite",
         "build": "vite build",
-        "preview": "vite preview"
+        "preview": "vite preview",
+        "lint": "eslint src",
+        "fix": "eslint src --fix"
     },
     "dependencies": {
         "axios": "^1.4.0",
@@ -16,13 +18,13 @@
         "vue-router": "^4.2.2"
     },
     "devDependencies": {
+        "@types/node": "^20.4.1",
         "@typescript-eslint/eslint-plugin": "^5.60.1",
         "@typescript-eslint/parser": "^5.60.1",
         "@vitejs/plugin-vue": "^4.2.3",
         "eslint": "^8.43.0",
         "eslint-plugin-vue": "^9.15.1",
         "less": "^4.1.3",
-        "less-loader": "^11.1.3",
         "prettier": "^2.8.8",
         "typescript": "^5.1.3",
         "vite": "^4.3.9"
