+++ web/src/views/login/index.vue
@@ -1,7 +1,15 @@
 <template>
-    <div>123456</div>
+    <div class="father">
+        <div class="son">123456</div>
+    </div>
 </template>
 
 <script setup lang="ts"></script>
 
-<style lang="less"></style>
+<style lang="less" scoped>
+.father {
+    .son {
+        font-size: 80px;
+    }
+}
+</style>
