+++ web/vite.config.ts
@@ -1,5 +1,5 @@
 /**
- * Copyright © 2023 OpenIM open source community. All rights reserved.
+ * Copyright © 2023 OpenIMSDK open source community. All rights reserved.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -14,16 +14,18 @@
  * limitations under the License.
  */
 
-import { createApp } from 'vue';
-import App from './App.vue';
-import axios from 'axios';
-import router from './router';
+import { fileURLToPath, URL } from 'node:url';
 
-// import { getSDK } from '@/utils/open-im-sdk-wasm';
+import { defineConfig } from 'vite';
+import vue from '@vitejs/plugin-vue';
 
-// const OpenIM = getSDK();
-// console.log(OpenIM);
-
-const app = createApp(App);
-app.config.globalProperties.$https = axios; // use axios
-app.use(router).mount('#app'); // mount the router on the app
+// https://vitejs.dev/config/
+export default defineConfig({
+    plugins: [vue()],
+    resolve: {
+        alias: {
+            '@': fileURLToPath(new URL('./src', import.meta.url)),
+        },
+        extensions: ['.js', '.ts', '.vue']
+    },
+});
